requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
crispy-bootstrap5==2024.10
Django==5.1.2
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-rosetta==0.10.1
idna==3.10
pillow==11.0.0
polib==1.2.0
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
